+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Filter - FlagValues within tags (#63)
+
 ## [1.11.0] - 2022-08-07
 
 - Added support for parameters in Cookies (#61)